Work on unit tests, school choice

+++ app/controllers/events_controller.rb
@@ -6,8 +6,10 @@ class EventsController < ApplicationController
   before_action :admin_user,          only: :destroy
 
   def index
-
-
+    if not user_signed_in?
+      redirect_to :signin
+    return
+    end
     @search = Sunspot.search(Event) do
       fulltext params[:search]
        with(:start).less_than(Time.zone.now)
@@ -57,38 +59,54 @@ class EventsController < ApplicationController
 #   end
 
   def show
-    respond_to do |format|
-      format.html do
-        @event = Event.find(params[:id])
+    if user_signed_in?
+      respond_to do |format|
+        format.html do
+          @event = Event.find(params[:id])
+        end
+        format.json do
+          @event = Event.find(params[:id])
+          render json: @event.as_json
+        end
       end
-      format.json do
-        @event = Event.find(params[:id])
-        render json: @event.as_json
-      end
+    else
+      redirect_to :signin
     end
   end
 
   def create
-    @event = current_user.events.build(event_params)
-    respond_to do |format|
-      format.html do
-        if @event.save
-          flash[:success] = 'Event created!'
-          redirect_to root_path
-        else
-          @feed_items = []
-          flash[:notice] = 'Event creation failed!'
-          render 'static_pages/home'
+    if user_signed_in?
+      @event = current_user.events.build(event_params)
+      respond_to do |format|
+        format.html do
+          if @event.save
+            flash[:success] = 'Event created!'
+            redirect_to root_path
+          else
+            @feed_items = []
+            flash[:notice] = 'Event creation failed!'
+            render 'static_pages/home'
+          end
+        end
+        format.json do
+          if @event.save
+            render :json => {:state => {:code => 0}, :data => @event.to_json }
+          else
+            @feed_items = []
+            render :json => {:state => {:code => 1, :messages => @event.errors.full_messages} }
+          end
         end
       end
-      format.json do
-        if @event.save
-          render :json => {:state => {:code => 0}, :data => @event.to_json }
-        else
-          @feed_items = []
-          render :json => {:state => {:code => 1, :messages => @event.errors.full_messages} }
-        end
-      end
+    else
+      redirect_to signin
+    end
+  end
+
+  def new
+    if user_signed_in?
+      @event = current_user.events.build
+    else
+      redirect_to :signin
     end
   end
 
@@ -97,37 +115,38 @@ class EventsController < ApplicationController
   end
 
   def update
+    params = event_params
+    @event = Event.find params[:id]
+
     respond_to do |format|
-      format.html do
-        @event = Event.find(params[:id])
-        if @event.update_attributes(event_params)
+      if @event && @event.update(params)
+        format.html do       
           flash[:success] = 'Event updated'
-          redirect_to root_path
-        else
-          render :edit, :alert => 'Unable to update event.'
+          redirect_to @event
         end
-      end
-      format.json do
-        @event = Event.find(params[:id])
-        if @event.update_attributes(event_params)
-          render :json => {:state => {:code => 0}, status: :ok, :data => @event.to_json }
-        else
-          render :json => {:state => {:code => 1, status: :error, :messages => @user.errors.full_messages} }
-        end
+        format.json {render :json => {:state => {:code => 0}, status: :ok, :data => @event.to_json }}
+        format.all { render_404 }
+      elsif @event != nil
+        format.html {render :edit, :alert => 'Unable to update event.'}
+        format.json {render :json => {:state => {:code => 1, status: :error, :messages => @user.errors.full_messages} }}
+        format.all {render_404}
       end
     end
   end
 
   def destroy
-    respond_to do |format|
-      format.html do
-        @event.destroy
-      end
-      format.json do
-        if @event.destroy
-          render :json => {:state => {:code => 0}}
-        else
-          render :json => {:state => {:code => 1, :messages => @user.errors.full_messages} }
+    if user_signed_in?
+      @event = Event.find params[:id]
+      respond_to do |format|
+        format.html do
+          @event.destroy
+        end
+        format.json do
+          if @event.destroy
+            render :json => {:state => {:code => 0}}
+          else
+            render :json => {:state => {:code => 1, :messages => @user.errors.full_messages} }
+          end
         end
       end
     end
@@ -156,6 +175,6 @@ class EventsController < ApplicationController
 
   # Confirms an admin user.
   def admin_user
-    redirect_to(root_url) unless current_user.role == 'Admin'
+    redirect_to(root_url) unless current_user && current_user.role == 'Admin'
   end
 end
